move runner into their own file
