+++ .gitignore
@@ -5,8 +5,6 @@
 *.so
 *.dylib
 
-draino
-
 # Test binary, build with `go test -c`
 *.test
